Add a tiny chart to bottom drawer toggler
